[FEAT] canvas list websocket 연결 분리 중
