+++ src/components/canvas/CanvasListComponent.vue
@@ -24,7 +24,8 @@
 
 <script>
 import axios from "axios";
-// import { useRouter } from "vue-router";
+import SockJS from "sockjs-client";
+import { Stomp } from "@stomp/stompjs";
 
 export default {
   name: "CanvasListComponent",
@@ -51,6 +52,11 @@ export default {
       canvasIdInList: null,
       channelId: null,
       chatrooms: [],
+
+      // socket 통신용
+
+      ws: null,
+      sock: null,
     };
   },
   methods: {
@@ -118,6 +124,82 @@ export default {
         }
       }
     },
+    connect() {
+      // 캔버스 소켓 연결
+      this.sock = new SockJS(`${process.env.VUE_APP_API_BASE_URL}/ws-stomp`);
+      this.ws = Stomp.over(this.sock);
+      this.ws.connect(
+        {},
+        () => {
+          this.ws.subscribe(
+            `/sub/canvas/room/${this.detailCanvasId}`,
+            (message) => {
+              const recv = JSON.parse(message.body);
+              this.recvMessage(recv);
+            }
+          );
+          this.ws.send(
+            `/pub/canvas/message`,
+            {},
+            JSON.stringify({
+              type: "ENTER",
+              roomId: this.detailCanvasId,
+              sender: this.sender,
+            })
+          );
+        },
+        () => {
+          if (this.reconnect++ <= 5) {
+            setTimeout(() => {
+              this.sock = new SockJS(
+                `${process.env.VUE_APP_API_BASE_URL}/ws-stomp`
+              );
+              this.ws = Stomp.over(this.sock);
+              this.connect();
+            }, 10 * 1000);
+          }
+        }
+      );
+    },
+    sendMessage() {
+      if (this.ws && this.ws.connected) {
+        this.ws.send(
+          `/pub/block/message`,
+          {},
+          JSON.stringify({
+            type: "CANVAS",
+            roomId: this.detailCanvasId,
+            sender: this.sender,
+            message: JSON.stringify(this.message),
+          })
+        );
+        this.message = "";
+      } else {
+        // console.log("WebSocket is not connected.");
+      }
+    },
+    recvMessage(recv) {
+      if (recv.type === "CANVAS") {
+        const canvasJson = JSON.parse(recv.message);
+        console.error("canvasJson >> ", canvasJson);
+      }
+    },
+    beforeRouteLeave() {
+      // 컴포넌트가 파괴되기 전에 구독 해제 및 WebSocket 연결 종료
+      if (this.sock) {
+        this.sock.close(); // SockJS 연결을 닫음
+        this.sock = null;
+        console.log("WebSocket subscription unsubscribed.");
+      }
+      if (this.ws) {
+        this.ws.disconnect(() => {
+          console.log("WebSocket connection closed.");
+        });
+      }
+    },
+  },
+  beforeUnmount() {
+    this.beforeRouteLeave();
   },
 };
 </script>
